just use a different data type

+++ src/protocols/dhcpv4/encode.c
@@ -145,14 +145,14 @@ static ssize_t encode_value(fr_dbuff_t *dbuff,
 		}
 
 		if (da_is_bits_prefix(vp->da)) {
-			int num_bytes = (vp->vp_ip.prefix + 0x07) >> 3;
+			size_t num_bytes = (vp->vp_ip.prefix + 0x07) >> 3;
 
 			fr_dbuff_in(&work_dbuff, (uint8_t) vp->vp_ip.prefix);
 
 			if (num_bytes) {
 				FR_DBUFF_IN_MEMCPY_RETURN(&work_dbuff,
 							  (uint8_t const *)&vp->vp_ipv4addr,
-							  (size_t) num_bytes);
+							  num_bytes);
 			}
 
 			break;
